refactor type definitions for extension and decoration handling
